cast value to int32 before exporting to avoid warning

+++ macros.h
@@ -71,7 +71,7 @@
 #define SN_EXPORT_UINT32(name, num) \
   { \
     napi_value name##_num; \
-    SN_STATUS_THROWS(napi_create_uint32(env, num, &name##_num), "") \
+    SN_STATUS_THROWS(napi_create_uint32(env, (int32_t) num, &name##_num), "") \
     SN_STATUS_THROWS(napi_set_named_property(env, exports, #name, name##_num), "") \
   }
